# packages/api-contract/openapi.json
{
  "components": {
    "schemas": {
...
      "CalibrationPolicy": {
        "properties": {
          "allow_retry": {
            "default": true,
            "title": "Allow Retry",
            "type": "boolean"
          },
          "maximum_attempts": {
            "default": 3,
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Maximum Attempts",
            "type": "integer"
          },
          "minimum_quality": {
            "default": "variable",
            "enum": [
              "strong",
              "variable",
              "limited"
            ],
            "title": "Minimum Quality",
            "type": "string"
          }
        },
        "title": "CalibrationPolicy",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "error",
          "request_id"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PublicStudyProtocol": {
        "properties": {
          "calibration_policy": {
            "$ref": "#/components/schemas/CalibrationPolicy"
          },
          "collection_policy": {
            "$ref": "#/components/schemas/CollectionPolicy"
          },
          "consent_text": {
            "title": "Consent Text",
            "type": "string"
          },
          "consent_version": {
            "title": "Consent Version",
            "type": "string"
          },
          "target_origins": {
            "items": {
              "format": "uri",
              "maxLength": 2083,
              "minLength": 1,
              "type": "string"
            },
            "title": "Target Origins",
            "type": "array"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/PublicTask"
            },
            "title": "Tasks",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "consent_version",
          "consent_text",
          "target_origins",
          "calibration_policy",
          "collection_policy",
          "tasks"
        ],
        "title": "PublicStudyProtocol",
        "type": "object"
      },
...
      "PublishResponse": {
        "properties": {
          "participant_link": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParticipantLinkResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "replayed": {
            "default": false,
            "title": "Replayed",
            "type": "boolean"
          },
          "study": {
            "$ref": "#/components/schemas/StudySummary"
          },
          "version": {
            "$ref": "#/components/schemas/StudyVersionResponse"
          }
        },
        "required": [
          "study",
          "version"
        ],
        "title": "PublishResponse",
        "type": "object"
      },
...
      "StudyListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StudySummary"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "StudyListResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Versioned REST API for research projects, published eye-tracking protocols, participant sessions, sample ingestion, and analysis.",
    "summary": "Eye-tracking research platform API",
    "title": "WebGaze API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/participate/{token}": {
      "get": {
        "operationId": "resolveParticipantLink",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicStudyProtocol"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Participant Link",
        "tags": [
          "studies"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/studies": {
      "get": {
        "operationId": "listStudies",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-demo-owner-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-Owner-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Studies",
        "tags": [
          "studies"
        ]
      },
      "post": {
        "operationId": "createStudy",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-demo-owner-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-Owner-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyDraftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Create Study",
        "tags": [
          "studies"
        ]
      }
    },
    "/api/v1/studies/{study_id}": {
      "delete": {
        "operationId": "deleteStudy",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Study Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-demo-owner-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-Owner-Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Study",
        "tags": [
          "studies"
        ]
      }
    },
    "/api/v1/studies/{study_id}/draft": {
      "get": {
        "operationId": "getStudyDraft",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Study Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-demo-owner-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-Owner-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyDraftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Study Draft",
        "tags": [
          "studies"
        ]
      },
      "put": {
        "operationId": "replaceStudyDraft",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Study Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-demo-owner-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-Owner-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyDraft"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyDraftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace Study Draft",
        "tags": [
          "studies"
        ]
      }
    },
    "/api/v1/studies/{study_id}/publish": {
      "post": {
        "operationId": "publishStudy",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Study Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "x-demo-owner-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-Owner-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish Study",
        "tags": [
          "studies"
        ]
      }
    },
    "/api/v1/studies/{study_id}/versions/{version_number}": {
      "get": {
        "operationId": "getStudyVersion",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Study Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version_number",
            "required": true,
            "schema": {
              "title": "Version Number",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-demo-owner-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-Owner-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyVersionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Study Version",
        "tags": [
          "studies"
        ]
      }
    },
...
  }
}